refactor: modelモジュールに移動
